WORKS - signup with email and password with auto sigIn and redirect - sigIn with google with redirect works - accoiunt linking works, when signing up with credentials firs, then logging in with google.
NEXT:
when a user tries logs in with email and password, but previously only signed up with the same email with provider google, there is an error showing.

I want to have this handled gracefully.
A user should be informed in a modal that he previously used google login, no email and password credentials are saved for this user, but can be added with the signup form.
The modal should have 2 buttons.
1) log in with google
2) SignUp with email and password with link to form.